Porting other tests and tweaking

+++ NdkBinderClient/src/main/cpp/tests/blackscholes/blackScholes.h
@@ -10,7 +10,7 @@ extern "C" {
 #endif
 
 //Main test function
-int startBalckScholes (bool perf);
+double startBalckScholes (bool perf);
 
 //Number of iterations in the test's main loop
 int getNewCount();
